Overhaul resetGUI fn. Prompt user when something goes sideways.
Prompt user to continue or not when (pallet not in DB and another pallet's
straws are on this ID) and when (co2 procedure for this pallet already exists).

resetGUI now kills the whole gui and re-opens it. So you have to log
back in and re-type the three fields. Sorry, but it was the most
straightforward way I could think of to clear the DP and session
completely. Was running into problems when I cleared it manually.